requirements.txt
python-dotenv==1.0.1
Flask==3.0.2
Flask-Caching==2.1.0
requests==2.31.0